requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==5.0.4
docutils==0.20.1
icalendar==5.0.11
idna==3.6
lockfile==0.12.2
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
schedule==1.2.1
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
urllib3==2.1.0
python-dotenv==1.0.1
